Fix: Cancel heartbeat after browser window reload (#1543)
* Fix: Cancel heartbeat after window reload

* Fix: Restore print after reconnect

* Fix: Fix auto reconnect after reload

+++ src/server/services/socket/socket-http.ts
@@ -198,18 +198,22 @@ class SocketHttp {
 
     public connectionClose = (socket: SocketServer, options: EventOptions) => {
         const { eventName } = options;
-        const api = `${this.host}/api/v1/disconnect`;
-        request
-            .post(api)
-            .timeout(3000)
-            .send(`token=${this.token}`)
-            .end((err, res) => {
-                socket && socket.emit(eventName, _getResult(err, res));
-            });
-        this.host = '';
-        this.token = '';
-        this.heartBeatWorker && this.heartBeatWorker.terminate();
-        clearInterval(intervalHandle);
+        if (this.host) {
+            const api = `${this.host}/api/v1/disconnect`;
+            request
+                .post(api)
+                .timeout(3000)
+                .send(`token=${this.token}`)
+                .end((err, res) => {
+                    socket && socket.emit(eventName, _getResult(err, res));
+                });
+            this.host = '';
+            this.token = '';
+            this.heartBeatWorker && this.heartBeatWorker.terminate();
+            clearInterval(intervalHandle);
+        } else {
+            socket && socket.emit(eventName, _getResult(new Error('connection not exist'), null));
+        }
     };
 
     public startGcode = (options: EventOptions) => {
@@ -405,10 +409,29 @@ class SocketHttp {
         request
             .get(api)
             .end((err, res) => {
-                this.socket && this.socket.emit(eventName, {
-                    msg: err?.message,
-                    text: res.text
-                });
+                if (err) {
+                    this.socket && this.socket.emit(eventName, {
+                        msg: err?.message,
+                        text: res.text
+                    });
+                } else {
+                    let gcodeStr = '';
+                    res.on('data', chunk => {
+                        gcodeStr += chunk;
+                    });
+                    res.once('end', () => {
+                        this.socket && this.socket.emit(eventName, {
+                            msg: err?.message,
+                            text: gcodeStr
+                        });
+                    });
+                    res.once('error', (error) => {
+                        this.socket && this.socket.emit(eventName, {
+                            msg: error?.message,
+                            text: ''
+                        });
+                    });
+                }
             });
     };
